Relation ManyToMany between user and role

+++ database/migrations/2020_08_20_075954_create_locations_table.php
@@ -21,7 +21,7 @@ class CreateLocationsTable extends Migration
             $table->string('address', 255);
         
             // FK
-           $table-> unsignedInteger('locality_id');
+           $table-> unsignedBigInteger('locality_id');
            $table->foreign('locality_id')->references('id')->on('localities')->onDelete('restrict')->onUpdate('cascade');
 
             $table->string('website', 255)->nullable();
